+++ tests/integration_tests/attr_expansion.rs
@@ -1,11 +1,11 @@
-use svgdx::transform_str;
+use svgdx::transform_str_default;
 
 #[test]
 fn test_expand_rect_xy_wh() {
     let input = r#"<rect xy="1 2" wh="3 4"/>"#;
     let expected = r#"<rect x="1" y="2" width="3" height="4"/>"#;
 
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 }
 
 #[test]
@@ -13,7 +13,7 @@ fn test_expand_xy1_xy2() {
     let input = r#"<line xy1="1 2" xy2="3 4"/>"#;
     let expected = r#"<line x1="1" y1="2" x2="3" y2="4"/>"#;
 
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 }
 
 #[test]
@@ -21,49 +21,49 @@ fn test_expand_rect_cxy_wh() {
     let input = r#"<rect cxy="5 7" wh="3 4"/>"#;
     let expected = r#"<rect x="3.5" y="5" width="3" height="4"/>"#;
 
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 }
 
 #[test]
 fn test_expand_rect_xy_loc() {
     let input = r#"<rect xy="5 7" wh="3 4" xy-loc="br"/>"#;
     let expected = r#"<rect x="2" y="3" width="3" height="4"/>"#;
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 
     let input = r#"<rect xy="5 7" wh="3 4" xy-loc="t"/>"#;
     let expected = r#"<rect x="3.5" y="7" width="3" height="4"/>"#;
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 
     let input = r#"<rect xy="5 7" wh="4 6" xy-loc="c"/>"#;
     let expected = r#"<rect x="3" y="4" width="4" height="6"/>"#;
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 }
 
 #[test]
 fn test_expand_cycle() {
     let input = r#"<rect xy="5.5" wh="2"/>"#;
     let expected = r#"<rect x="5.5" y="5.5" width="2" height="2"/>"#;
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 }
 
 #[test]
 fn test_expand_circle() {
     let input = r#"<circle cxy="5" wh="3"/>"#;
     let expected = r#"<circle cx="5" cy="5" r="1.5"/>"#;
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 
     let input = r#"<circle xy="5" wh="3"/>"#;
     let expected = r#"<circle cx="6.5" cy="6.5" r="1.5"/>"#;
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 }
 
 #[test]
 fn test_expand_ellipse() {
     let input = r#"<ellipse cxy="5" wh="3 2"/>"#;
     let expected = r#"<ellipse cx="5" cy="5" rx="1.5" ry="1"/>"#;
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 
     let input = r#"<ellipse xy="5" wh="3 2"/>"#;
     let expected = r#"<ellipse cx="6.5" cy="6" rx="1.5" ry="1"/>"#;
-    assert_eq!(transform_str(input).unwrap(), expected);
+    assert_eq!(transform_str_default(input).unwrap(), expected);
 }