+++ tests/integration_tests/rel_pos.rs
@@ -1,4 +1,4 @@
-use svgdx::transform_str;
+use svgdx::transform_str_default;
 
 #[test]
 fn test_rel_prev() {
@@ -7,14 +7,18 @@ fn test_rel_prev() {
 <rect xy="@tr" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="30" y="20" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 
     let rel_v_input = r#"
 <rect xy="10 20" wh="20 60" />
 <rect xy="@bl -1 1" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="9" y="81" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -26,7 +30,7 @@ fn test_rel_refid() {
 <rect xy="#abc@tr" wh="20 60" id="z"/>
 "##;
     let expected_rect = r#"<rect x="30" y="20" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_refid_input)
+    assert!(transform_str_default(rel_refid_input)
         .unwrap()
         .contains(expected_rect));
 }
@@ -39,7 +43,9 @@ fn test_relh() {
 <rect xy="^h" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="30" y="20" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // With a gap
     let rel_h_input = r#"
@@ -47,7 +53,9 @@ fn test_relh() {
 <rect xy="^h 3" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="33" y="20" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // TO THE LEFT
     let rel_h_input = r#"
@@ -55,7 +63,9 @@ fn test_relh() {
 <rect xy="^H" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="-10" y="20" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // With a gap
     let rel_h_input = r#"
@@ -63,7 +73,9 @@ fn test_relh() {
 <rect xy="^H 3" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="-13" y="20" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -74,7 +86,9 @@ fn test_relv() {
 <rect xy="^v" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="10" y="80" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // With a gap
     let rel_v_input = r#"
@@ -82,7 +96,9 @@ fn test_relv() {
 <rect xy="^v 5" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="10" y="85" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // VERTICAL ABOVE
     let rel_v_input = r#"
@@ -90,7 +106,9 @@ fn test_relv() {
 <rect xy="^V" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="10" y="-40" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // With a gap
     let rel_v_input = r#"
@@ -98,7 +116,9 @@ fn test_relv() {
 <rect xy="^V 5" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="10" y="-45" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -110,7 +130,9 @@ fn test_elref_relh() {
 <rect id="z" xy="#abc:h" wh="10" />
 "##;
     let expected_rect = r#"<rect id="z" x="20" y="20" width="10" height="10"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // with a gap
     let rel_h_input = r##"
@@ -119,7 +141,9 @@ fn test_elref_relh() {
 <rect id="z" xy="#abc:h 3" wh="10" />
 "##;
     let expected_rect = r#"<rect id="z" x="23" y="20" width="10" height="10"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // TO THE LEFT
     let rel_h_input = r##"
@@ -128,7 +152,9 @@ fn test_elref_relh() {
 <rect id="z" xy="#abc:H" wh="8" />
 "##;
     let expected_rect = r#"<rect id="z" x="2" y="21" width="8" height="8"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // with a gap
     let rel_h_input = r##"
@@ -137,7 +163,9 @@ fn test_elref_relh() {
 <rect id="z" xy="#abc:H 3" wh="8" />
 "##;
     let expected_rect = r#"<rect id="z" x="-1" y="21" width="8" height="8"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -149,7 +177,9 @@ fn test_elref_relv() {
 <rect id="z" xy="#abc:v" wh="10" />
 "##;
     let expected_rect = r#"<rect id="z" x="10" y="30" width="10" height="10"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // with a gap
     let rel_v_input = r##"
@@ -158,7 +188,9 @@ fn test_elref_relv() {
 <rect id="z" xy="#abc:v 3" wh="10" />
 "##;
     let expected_rect = r#"<rect id="z" x="10" y="33" width="10" height="10"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // ABOVE
     let rel_v_input = r##"
@@ -167,7 +199,9 @@ fn test_elref_relv() {
 <rect id="z" xy="#abc:V" wh="8" />
 "##;
     let expected_rect = r#"<rect id="z" x="11" y="12" width="8" height="8"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 
     // with a gap
     let rel_v_input = r##"
@@ -176,7 +210,9 @@ fn test_elref_relv() {
 <rect id="z" xy="#abc:V 3" wh="8" />
 "##;
     let expected_rect = r#"<rect id="z" x="11" y="9" width="8" height="8"/>"#;
-    assert!(transform_str(rel_v_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_v_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -186,7 +222,9 @@ fn test_rel_dx_dy() {
 <rect xy="^h" dxy="-1.23 4.56" wh="20 60" id="z"/>
 "#;
     let expected_rect = r#"<rect x="28.77" y="24.56" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_h_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_h_input)
+        .unwrap()
+        .contains(expected_rect));
 
     let rel_input = r##"
 <rect xy="10 20" wh="20 60" id="abc"/>
@@ -195,7 +233,9 @@ fn test_rel_dx_dy() {
 <rect xy="#abc@tr 10 100" wh="20 60" id="z"/>
 "##;
     let expected_rect = r#"<rect x="40" y="120" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -214,15 +254,15 @@ fn test_rel_recursive() {
 <rect xy="@tl 1 1" wh="2" id="z"/>
 "##;
     let expected_rect = r#"<rect x="42" y="20" width="10" height="10" id="x"/>"#;
-    assert!(transform_str(rel_refid_input)
+    assert!(transform_str_default(rel_refid_input)
         .unwrap()
         .contains(expected_rect));
     let expected_rect = r#"<rect x="64" y="20" width="7" height="7" id="y"/>"#;
-    assert!(transform_str(rel_refid_input)
+    assert!(transform_str_default(rel_refid_input)
         .unwrap()
         .contains(expected_rect));
     let expected_rect = r#"<rect x="67" y="23" width="2" height="2" id="z"/>"#;
-    assert!(transform_str(rel_refid_input)
+    assert!(transform_str_default(rel_refid_input)
         .unwrap()
         .contains(expected_rect));
 }