+++ tests/integration_tests/rel_size.rs
@@ -1,4 +1,4 @@
-use svgdx::transform_str;
+use svgdx::transform_str_default;
 
 #[test]
 fn test_rel_size_prev() {
@@ -7,7 +7,9 @@ fn test_rel_size_prev() {
 <rect xy="25 35" wh="^" id="z"/>
 "#;
     let expected_rect = r#"<rect x="25" y="35" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_wh_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_wh_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -19,7 +21,7 @@ fn test_rel_size_refid() {
 <rect xy="12 0" wh="#abc" id="z"/>
 "##;
     let expected_rect = r#"<rect x="12" y="0" width="20" height="60" id="z"/>"#;
-    assert!(transform_str(rel_size_refid_input)
+    assert!(transform_str_default(rel_size_refid_input)
         .unwrap()
         .contains(expected_rect));
 }
@@ -31,7 +33,7 @@ fn test_rel_size_dxy() {
 <rect xy="1 2" wh="^ 2 -5" id="z"/>
 "#;
     let expected_rect = r#"<rect x="1" y="2" width="22" height="55" id="z"/>"#;
-    assert!(transform_str(rel_size_dxy_input)
+    assert!(transform_str_default(rel_size_dxy_input)
         .unwrap()
         .contains(expected_rect));
 
@@ -42,7 +44,9 @@ fn test_rel_size_dxy() {
 <rect xy="1 2" wh="#abc -2 5" id="z"/>
 "##;
     let expected_rect = r#"<rect x="1" y="2" width="18" height="65" id="z"/>"#;
-    assert!(transform_str(rel_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -52,7 +56,7 @@ fn test_rel_size_dxy_pct() {
 <rect xy="1 2" wh="^ 110% 50%" id="z"/>
 "#;
     let expected_rect = r#"<rect x="1" y="2" width="22" height="30" id="z"/>"#;
-    assert!(transform_str(rel_size_dxy_pct_input)
+    assert!(transform_str_default(rel_size_dxy_pct_input)
         .unwrap()
         .contains(expected_rect));
 
@@ -63,7 +67,9 @@ fn test_rel_size_dxy_pct() {
 <rect xy="1 2" wh="#abc 40% 150%" id="z"/>
 "##;
     let expected_rect = r#"<rect x="1" y="2" width="8" height="90" id="z"/>"#;
-    assert!(transform_str(rel_input).unwrap().contains(expected_rect));
+    assert!(transform_str_default(rel_input)
+        .unwrap()
+        .contains(expected_rect));
 }
 
 #[test]
@@ -82,15 +88,15 @@ fn test_rel_size_recursive() {
 <rect xy="1 1" wh="#y" id="z"/>
 "##;
     let expected_rect = r#"<rect x="12" y="34" width="23" height="30" id="x"/>"#;
-    assert!(transform_str(rel_recur_input)
+    assert!(transform_str_default(rel_recur_input)
         .unwrap()
         .contains(expected_rect));
     let expected_rect = r#"<rect x="2" y="2" width="46" height="27" id="y"/>"#;
-    assert!(transform_str(rel_recur_input)
+    assert!(transform_str_default(rel_recur_input)
         .unwrap()
         .contains(expected_rect));
     let expected_rect = r#"<rect x="1" y="1" width="46" height="27" id="z"/>"#;
-    assert!(transform_str(rel_recur_input)
+    assert!(transform_str_default(rel_recur_input)
         .unwrap()
         .contains(expected_rect));
 }