+++ tests/integration_tests/repeat_el.rs
@@ -1,4 +1,4 @@
-use svgdx::transform_str;
+use svgdx::transform_str_default;
 
 #[test]
 fn test_repeat_simple() {
@@ -12,7 +12,10 @@ fn test_repeat_simple() {
 <rect x="4" y="0" width="2" height="2"/>
 <rect x="6" y="0" width="2" height="2"/>
 "#;
-    assert_eq!(transform_str(rep_input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(rep_input).unwrap().trim(),
+        expected.trim()
+    );
 }
 
 #[test]
@@ -24,5 +27,8 @@ fn test_repeat_zero() {
     let expected = r#"
 <rect x="0" y="0" width="2" height="2"/>
 "#;
-    assert_eq!(transform_str(rep_input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(rep_input).unwrap().trim(),
+        expected.trim()
+    );
 }