+++ tests/integration_tests/text_attr.rs
@@ -1,4 +1,4 @@
-use svgdx::transform_str;
+use svgdx::transform_str_default;
 
 #[test]
 fn test_basic_rect_text() {
@@ -10,7 +10,10 @@ fn test_basic_rect_text() {
 <text x="2.5" y="3" class="d-tbox">thing</text>
 "#;
 
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 }
 
 #[test]
@@ -23,7 +26,10 @@ fn test_expanded_rect_text() {
 <text x="20" y="20" class="d-tbox">thing</text>
 "#;
 
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 }
 
 #[test]
@@ -36,7 +42,10 @@ fn test_text_loc() {
 <text x="20" y="11" class="d-tbox d-text-top">thing</text>
 "#;
 
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 
     let input = r#"
 <rect cxy="20" wh="20" text="thing" text-loc="bl"/>
@@ -46,7 +55,10 @@ fn test_text_loc() {
 <text x="11" y="29" class="d-tbox d-text-bottom d-text-left">thing</text>
 "#;
 
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 }
 
 #[test]
@@ -60,7 +72,10 @@ fn test_text_multiline() {
 <tspan x="5" dy="-0.525em">multi</tspan><tspan x="5" dy="1.05em">line</tspan>
 </text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 
     let input = r#"
 <rect xy="0" wh="10" text-loc="t" text-lsp="2" text="multi\nline"/>
@@ -71,7 +86,10 @@ fn test_text_multiline() {
 <tspan x="5" dy="0em">multi</tspan><tspan x="5" dy="2em">line</tspan>
 </text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 
     let input = r#"
 <rect xy="0" wh="10" text-loc="br" text-lsp="1" text="multi\nline"/>
@@ -82,7 +100,10 @@ fn test_text_multiline() {
 <tspan x="9" dy="-1em">multi</tspan><tspan x="9" dy="1em">line</tspan>
 </text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 
     // test adjacent empty lines
     let input = r#"
@@ -94,7 +115,10 @@ fn test_text_multiline() {
 <tspan x="5" dy="-1.5em">multi</tspan><tspan x="5" dy="1em"></tspan><tspan x="5" dy="1em"></tspan><tspan x="5" dy="1em">line</tspan>
 </text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 }
 
 #[test]
@@ -106,7 +130,10 @@ fn test_text_dxy() {
 <rect x="0" y="0" width="10" height="10"/>
 <text x="7" y="5" class="d-tbox">blob</text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 
     let input = r#"
 <rect xy="0" wh="10" text-dy="-2" text="blob"/>
@@ -115,7 +142,10 @@ fn test_text_dxy() {
 <rect x="0" y="0" width="10" height="10"/>
 <text x="5" y="3" class="d-tbox">blob</text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 
     let input = r#"
 <rect xy="0" wh="10" text-dxy="1.5 3" text="blob"/>
@@ -124,7 +154,10 @@ fn test_text_dxy() {
 <rect x="0" y="0" width="10" height="10"/>
 <text x="6.5" y="8" class="d-tbox">blob</text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 }
 
 #[test]
@@ -138,7 +171,10 @@ fn test_text_line() {
 <line x1="0" y1="0" x2="10" y2="0"/>
 <text x="5" y="0" class="d-tbox">blob</text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 
     let input = r#"
 <line xy1="0" xy2="10 0" text="blob" text-loc="r"/>
@@ -147,5 +183,8 @@ fn test_text_line() {
 <line x1="0" y1="0" x2="10" y2="0"/>
 <text x="11" y="0" class="d-tbox d-text-left">blob</text>
 "#;
-    assert_eq!(transform_str(input).unwrap().trim(), expected.trim());
+    assert_eq!(
+        transform_str_default(input).unwrap().trim(),
+        expected.trim()
+    );
 }